source: outline `chunks` from `model`
